+++ agents/src/job.ts
@@ -211,15 +211,15 @@ export class JobContext {
   /** @internal */
   onParticipantConnected(p: RemoteParticipant) {
     for (const callback of this.#participantEntrypoints) {
-      if (this.#participantTasks[p.identity]?.callback == callback) {
+      if (this.#participantTasks[p.identity!]?.callback == callback) {
         this.#logger.warn(
           'a participant has joined before a prior prticipant task matching the same identity has finished:',
           p.identity,
         );
       }
       const result = callback(this, p);
-      result.finally(() => delete this.#participantTasks[p.identity]);
-      this.#participantTasks[p.identity] = { callback, result };
+      result.finally(() => delete this.#participantTasks[p.identity!]);
+      this.#participantTasks[p.identity!] = { callback, result };
     }
   }